BACKPORT: fuse: allow using readdir cache
The cache is only used if it's completed, not while it's still being
filled; this constraint could be lifted later, if it turns out to be
useful.

Introduce state in struct fuse_file that indicates the position within the
cache.  After a seek, reset the position to the beginning of the cache and
search the cache for the current position.  If the current position is not
found in the cache, then fall back to uncached readdir.

It can also happen that page(s) disappear from the cache, in which case we
must also fall back to uncached readdir.

Change-Id: I4ec64ae5f6b48b6f2ca3bfe6b01747329d110844
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Kneba <kneba@users.noreply.github.com>