UPSTREAM: selinux: Remove SECURITY_SELINUX_BOOTPARAM_VALUE
In preparation for removing CONFIG_DEFAULT_SECURITY, this removes the
soon-to-be redundant SECURITY_SELINUX_BOOTPARAM_VALUE. Since explicit
ordering via CONFIG_LSM or "lsm=" will define whether an LSM is enabled or
not, this CONFIG will become effectively ignored, so remove it. However,
in order to stay backward-compatible with "security=selinux", the enable
variable defaults to true.

Change-Id: I6c02365f15d135a63857e5bebccee4654ed5c39e
Signed-off-by: Kees Cook <keescook@chromium.org>
Signed-off-by: Kneba <kneba@users.noreply.github.com>