UPSTREAM: fuse: extract helper for range writeback
The fuse_writeback_range() helper flushes dirty data to the userspace
filesystem.

When the function returns, the WRITE requests for the data in the given
range have all been completed.  This is not equivalent to fsync() on the
given range, since the userspace filesystem may not yet have the data on
stable storage.

Change-Id: I192f8fb55caa066e7d8be61dfb0fd0f7f220abf9
Signed-off-by: Miklos Szeredi <mszeredi@redhat.com>
Signed-off-by: Kneba <kneba@users.noreply.github.com>