simple_lmk: Consider all positive adjs when finding victims
We are allowed to kill any process with a positive adj, so we shouldn't
exclude any processes with adjs greater than 999. This would present a
problem with quirky applications that set their own adj score, such as
stress-ng. In the case of stress-ng, it would set its adj score to 1000
and thus exempt itself from being killed by Simple LMK. This shouldn't
be allowed; any process with a positive adj, up to the highest positive
adj possible (32767) should be killable.

Reported-by: Danny Lin <danny@kdrag0n.dev>
Change-Id: I6b66d01a62b8edce18127ad122a59907f128115f

Signed-off-by: RyuujiX <saputradenny712@gmail.com>