qcacld-3.0: Update channel self Rx time to the user space
Currently, channel rx time is not being updated to user space as
part of link stats. Instead channel tx_time is being overwritten
with the same value. Due to this user space is not getting channel
rx_time in the link stats.

Update channel rx time in the channel link stats.

Update representation of tx_time and rx_time in the ll_stats debugfs
output to align with other stats.

Change-Id: I850adb11169eb24c81f1e31a4ff498f4e20ab9c6
CRs-Fixed: 2765485
Signed-off-by: RyuujiX <saputradenny712@gmail.com>