bdi: remove BDI_CAP_SYNCHRONOUS_IO
BDI_CAP_SYNCHRONOUS_IO is only checked in the swap code, and used to
decided if ->rw_page can be used on a block device.  Just check up for
the method instead.  The only complication is that zram needs a second
set of block_device_operations as it can switch between modes that
actually support ->rw_page and those who don't.

Signed-off-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Jan Kara <jack@suse.cz>
Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>
Change-Id: I76402462142f15e2d90cc57d346b17ded34a97f6
Signed-off-by: UtsavBalar1231 <utsavbalar1231@gmail.com>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>