simple_lmk: Be extra paranoid if tasks can have no pages
If it's possible for a task to have no pages, then there could be a case
where `pages_found` is zero while `nr_found` isn't, which would cause
the found tasks' locks to never be unlocked, and thus mayhem. We can
change the `pages_found` check to use `nr_found` instead in order to
naturally defend against this scenario, in case it is indeed possible.

Change-Id: I3ebd8442fc0218af418a7a05950257f01d626ca3

Signed-off-by: RyuujiX <saputradenny712@gmail.com>