Regulator: cpr3: fix -Wmisleading-indentation
Fixes:
drivers/regulator/cpr3-regulator.c:6491:3: error: misleading
indentation; statement is not part of the previous 'if'
[-Werror,-Wmisleading-indentation]
if (rc)
^
drivers/regulator/cpr3-regulator.c:6489:2: note: previous statement is here
if (ctrl->ctrl_type == CPR_CTRL_TYPE_CPR4)
^

Slightly more interesting case.  rc is initialized to zero, so it
doesn't matter if this condition is within the parent condition or not.
Do less work and move it into the condition as it was likely originally
intended.

Bug: 155426751
Change-Id: Iac4bf04f5b9aec59dd74028970dc111e48fa9e2a