techpack: audio: Reduce verbosity of logging
These are just useless. There are some logs that show up in dmesg
meaning that there are some bugs introduced by CAF's techpack.
At the end, everything works. Demote the useless logs to debug including
the ones that let us know about the bug until we find a fix to them, as
they are quite spammy.

Change-Id: Ib7871542f35f4c9d645176e6018eafc7e74cfc3b
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>