usb: gadget: configfs: Fix mismerge in string handling
Commit c3a4fc9f37af3ae ("usb: gadget: configfs: Fix KASAN use-after-free") has resolved
a use-after-free issue that made use of strcpy() but accidentally it was never imported,
resulting in breaking the original functionality of the patch.

strlcpy() is not suitable here since it may result in truncated strings, causing potential
issues with memory access and data integrity. Therefore, replace it with the originally
intended strcpy().

Fixes: a32dee2fa2874 ("ACK: Merge tag `v4.19.252` of android-4.19-stable")
Change-Id: If619e8d9aae6c6faab4973c8e323c08ca8729eb1
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>